// src/pages/Positions/types/iibbTypes.ts
export interface IIBBData {
  concepto: string;
  import: number;
  distrimar: number;
  junimar: number;
  gondolaLed: number;
  warnesTelas: number;
  aladmar: number;
  aftermarket: number;
}

export interface PerceptionData {
  concepto: string;
  import: number;
  distrimar: number;
  junimar: number;
}